Refactor backend into feature modules (profile, opportunity, feed, messaging, connection, user)
Replace flat controllers/dto/models/mappers/repository/service layout with bounded packages under com.agora.agoracampus. Add feed (posts, comments, reactions), messaging, and connection domains. Align competition and internship not-found exceptions with the new structure.

Made-with: Cursor

+++ AgoraCampus/src/main/java/com/agora/agoracampus/exception/CompetitionNotFoundException.java
@@ -2,7 +2,7 @@ package com.agora.agoracampus.exception;
 
 public class CompetitionNotFoundException extends RuntimeException {
 
-    public CompetitionNotFoundException(Integer id) {
+    public CompetitionNotFoundException(Long id) {
         super("Competitia cu id-ul " + id + " nu a fost gasita.");
     }
 }
