+++ AgoraCampus/src/main/java/com/agora/agoracampus/exception/InternshipNotFoundException.java
@@ -2,7 +2,7 @@ package com.agora.agoracampus.exception;
 
 public class InternshipNotFoundException extends RuntimeException {
 
-    public InternshipNotFoundException(Integer id) {
+    public InternshipNotFoundException(Long id) {
         super("Internship-ul cu id-ul " + id + " nu a fost gasit.");
     }
 }
